refactor(driver): gate native runtimes with explicit tags
