+++ pkg/driver/boxlite_image_materialize_cgo.go
@@ -1,4 +1,4 @@
-//go:build boxlitecgo
+//go:build linux && cgo && boxlitecgo
 
 package driver
