+++ pkg/driver/local_docker_oci_test.go
@@ -1,4 +1,4 @@
-//go:build cgo
+//go:build linux && cgo && (boxlitecgo || microsandboxcgo)
 
 package driver
