+++ pkg/driver/microsandbox_image_resolver.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo && microsandboxcgo
+
 package driver
 
 import (
