+++ pkg/driver/microsandbox_image_resolver_test.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo && microsandboxcgo
+
 package driver
 
 import (
