+++ pkg/driver/microsandbox_stats_test.go
@@ -1,4 +1,4 @@
-//go:build cgo
+//go:build linux && cgo && microsandboxcgo
 
 package driver
