+++ pkg/driver/runtime_cache_sources_nomicrosandbox.go
@@ -1,4 +1,4 @@
-//go:build !cgo
+//go:build !linux || !cgo || !microsandboxcgo
 
 package driver
