+++ pkg/driver/runtime_mount_manifest_boxlite_smoke_test.go
@@ -1,4 +1,4 @@
-//go:build boxlitecgo
+//go:build linux && cgo && boxlitecgo
 
 package driver
