+++ pkg/driver/runtime_mount_manifest_microsandbox_smoke_test.go
@@ -1,4 +1,4 @@
-//go:build cgo
+//go:build linux && cgo && microsandboxcgo
 
 package driver
