+++ pkg/driver/runtime_mount_manifest_smoke_test.go
@@ -1,4 +1,4 @@
-//go:build boxlitecgo || cgo
+//go:build linux && cgo && (boxlitecgo || microsandboxcgo)
 
 package driver
