+++ pkg/driver/runtime_oci_image_smoke_test.go
@@ -1,3 +1,5 @@
+//go:build linux && cgo && (boxlitecgo || microsandboxcgo)
+
 package driver
 
 import (
